try to fix overflow

+++ source/api_cc/src/commonPT.cc
@@ -3,12 +3,12 @@
 #include "commonPT.h"
 using namespace deepmd;
 void NeighborListDataPT::copy_from_nlist(const InputNlist& inlist,
-                                         int& max_num_neighbors) {
+                                         unsigned long int& max_num_neighbors) {
   int inum = inlist.inum;
   ilist.resize(inum);
   numneigh.resize(inum);
   memcpy(&ilist[0], inlist.ilist, inum * sizeof(int));
-  int* max_element = std::max_element(inlist.numneigh, inlist.numneigh + inum);
+  unsigned long int* max_element = std::max_element(inlist.numneigh, inlist.numneigh + inum);
   max_num_neighbors = *max_element;
   jlist.resize(inum * max_num_neighbors);
   memset(&jlist[0], -1, inum * max_num_neighbors * sizeof(int));
